requirements.txt
click==7.1.2
cycler==0.10.0
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==3.0.0a1
kiwisolver==1.2.0
MarkupSafe==2.0.0a1
numpy==1.18.4
pandas==1.0.3
plotly==4.7.0
pyparsing==2.4.7
pyserial==3.4
python-dateutil==2.8.1
pytz==2020.1
retrying==1.3.3
six==1.14.0
Werkzeug==1.0.1
